feat: Implementação Aula 03/10/2023

+++ backend/src/controllers/topic.controller.ts
@@ -1,4 +1,4 @@
-import { Controller,  Get,  Param,  Body,  ParseIntPipe,  Post,  Delete, HttpCode} from '@nestjs/common';
+import { Controller, Get, Param, Body, ParseIntPipe, Post, Delete, HttpCode, HttpException, HttpStatus, Put } from '@nestjs/common';
 import { Topic } from 'src/entities/topic.entity';
 import { TopicService } from 'src/services/topic.service';
 
@@ -12,18 +12,41 @@ export class TopicController {
   }
 
   @Get(':id')
-  findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
-    return this.service.findById(id);
+  async findById(@Param('id', ParseIntPipe) id: number): Promise<Topic> {
+    const found = await this.service.findById(id);
+
+    if (!found) {
+      throw new HttpException('Topic not found', HttpStatus.NOT_FOUND);
+    }
+
+    return found;
   }
 
   @Post()
-  create(@Body() Topic: Topic): Promise<Topic> {
-    return this.service.create(Topic);
+  create(@Body() topic: Topic): Promise<Topic> {
+    return this.service.create(topic);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
-    return this.service.delete(id);
+  async delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
+    const found = await this.service.findById(id);
+
+    if (!found) {
+      throw new HttpException('Topic not found', HttpStatus.NOT_FOUND);
+    }
+
+    await this.service.delete(id);
+  }
+
+  @Put(':id')
+  async update(@Param('id', ParseIntPipe) id: number, @Body() topic: Topic): Promise<Topic> {
+    const found = await this.service.findById(id);
+
+    if (!found) {
+      throw new HttpException('Topic not found', HttpStatus.NOT_FOUND);
+    }
+
+    return this.service.update(id, topic);
   }
 }
