+++ backend/src/controllers/user.controller.ts
@@ -1,6 +1,19 @@
-import { Controller,  Get,  Param,  Body,  ParseIntPipe,  Post,  Delete, HttpCode} from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Param,
+  Body,
+  ParseIntPipe,
+  Post,
+  Delete,
+  HttpCode,
+  HttpException,
+  HttpStatus,
+  Put,
+} from '@nestjs/common';
 import { User } from 'src/entities/user.entity';
 import { UserService } from 'src/services/user.service';
+import { Repository } from 'typeorm';
 
 @Controller('users')
 export class UserController {
@@ -12,8 +25,25 @@ export class UserController {
   }
 
   @Get(':id')
-  findById(@Param('id', ParseIntPipe) id: number): Promise<User> {
-    return this.service.findById(id);
+  async findById(@Param('id', ParseIntPipe) id: number): Promise<User> {
+    const found = await this.service.findById(id);
+
+    if (!found) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return found;
+  }
+
+  @Get(':usename')
+  async findByUsername(@Param('usename') usename: string): Promise<User> {
+    const found = await this.service.findByUsername(usename);
+
+    if (!found) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return found;
   }
 
   @Post()
@@ -23,7 +53,27 @@ export class UserController {
 
   @Delete(':id')
   @HttpCode(204)
-  delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
-    return this.service.delete(id);
+  async delete(@Param('id', ParseIntPipe) id: number): Promise<void> {
+    const found = await this.service.findById(id);
+
+    if (!found) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    await this.service.delete(id);
+  }
+
+  @Put(':id')
+  async update(
+    @Param('id', ParseIntPipe) id: number,
+    @Body() user: User,
+  ): Promise<User> {
+    const found = await this.service.findById(id);
+
+    if (!found) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return this.service.update(id, user);
   }
 }
